Improve ScreenSpace, hashTo5/6Bits, and ChallengeSortNode::SetNewIcon matches
ScreenSpace: 27.1% to 92.6% (fix rect.w/h swap, int to float casts, fsel)
hashTo5Bits: 77.8% to 98.2% (bool intermediate for Size()>2)
hashTo6Bits: 77.8% to 98.2% (same pattern)
ChallengeSortNode::SetNewIcon: 97.3% to 98.1% (signed compare cast)

+++ src/lazer/meta_ham/ChallengeSortNode.cpp
@@ -206,7 +206,7 @@ void ChallengeSortNode::SetNewIcon(UILabel *label) const {
     int timestamp = TheChallengeSortMgr->GetOwnerChallengeTimeStamp(
         mChallengeRecord->GetChallengeRow().mSongID
     );
-    if (timestamp > mChallengeRecord->GetChallengeRow().mTimeStamp
+    if (timestamp > (int)mChallengeRecord->GetChallengeRow().mTimeStamp
         || mChallengeRecord->GetUnk48() == mChallengeRecord->GetUnk4c()
         || mChallengeRecord->GetUnk50() == 4 || mChallengeRecord->GetUnk50() == 2
         || mChallengeRecord->GetUnk50() == 3) {
